backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.4
pandas==3.0.2
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
